feat: 최고 점수에 체력 조절 로직 추가
- 이제 적은 BestScore에 따라 체력 강화가 이루어져서 등장함

+++ Assets/02.Scripts/Projectile/Bomb.cs
@@ -6,7 +6,7 @@ public class Bomb : MonoBehaviour
     [SerializeField] private float _maxLifeTime = 10f;
     private float _lifeTime = 0f;
     [SerializeField] private LayerMask _damageableMask;
-    [SerializeField] private float _damage;
+    [SerializeField] private int _damage;
 
     [Header("Effect Setting")]
     [SerializeField] private Transform _trailTransform;
